requirements.txt
fastapi
fonttools
matplotlib
numpy
pillow
python-dateutil